(PE-34952) Add delete action with --expired flag
This adds the puppetserver ca delete action, which deletes signed certs on disk. While all the flags are defined here, only --expired is implemented so far. This searches the inventory file for any expired certs, deletes them, then finds any certs that are not in the inventory file, checks if they are expired, and deletes those.